requirements.txt
absl-py==2.1.0
astropy==6.1.4
astropy-iers-data==0.2024.10.28.0.34.7
astunparse==1.6.3
attrs==24.2.0
bilby==2.3.0
bilby.cython==0.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
corner==2.2.2
cryptography==43.0.3
cycler==0.12.1
dateparser==1.2.0
dill==0.3.9
dqsegdb2==1.2.1
dynesty==2.1.4
emcee==3.1.6
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
gwdatafind==1.2.0
gwosc==0.7.1
gwpy==3.0.10
h5py==3.12.1
idna==3.10
igwn-auth-utils==1.1.1
keras==3.6.0
kiwisolver==1.4.7
lalsuite==7.23
libclang==18.1.1
-e git+https://git.ligo.org/lscsoft/ligo-segments.git@be7c93b1ef431908a9df0f078815ee7f00345088#egg=ligo_segments
ligotimegps==2.0.1
lscsoft-glue==4.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nestle==0.2.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pycparser==2.22
pyerfa==2.0.1.4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.3
safe-netrc==1.0.1
scipy==1.14.1
scitokens==1.8.1
setuptools==75.3.0
six==1.16.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.16.0
